Animal page 재 로그인 문구 이동
