+++ tomyanimal/src/Animals/Checklist/AddChecklist.js
@@ -9,7 +9,6 @@ const AddChecklist = ( {submitHandler} ) => {
 
     const handleChange = (e)=> {
         if(characterLimit - e.target.value.length >= 0) {
-            //setTitle(e.target.value);
             setContent(e.target.value);
         }
     }
@@ -33,7 +32,6 @@ const AddChecklist = ( {submitHandler} ) => {
             placeholder='title'
             value={title}
             ref={contentRef}
-            //onChange={handleChange}
             onChange={(e) => setTitle(e.target.value)}
         />
         <textarea 
@@ -44,14 +42,12 @@ const AddChecklist = ( {submitHandler} ) => {
             value={content}
             ref={contentRef}
             onChange={handleChange}
-            //onChange={(e) => setContent(e.target.value)}
         />
         <div className='checklist__note__footer'>
             <small>{characterLimit - content.length} 글자</small>
             <button 
                 className='checklist__save'
                 onClick={handleSave}
-                //onClick={submitHandler}
             >
                 Save
             </button>
